cleanup: split capture.c into 2 separate files
